Fix docs and regenerate scaffolding. (#5)
