+++ .cloud-repo-tools.json
@@ -2,23 +2,21 @@
   "requiresKeyFile": true,
   "requiresProjectId": true,
   "product": "logging",
-  "client_reference_url": 
+  "client_reference_url":
     "https://cloud.google.com/logging/docs/setup/nodejs",
   "release_quality": "beta",
   "samples": [
     {
-      "id": "logs",
-      "name": "Logging",
-      "file": "logs.js",
-      "docs_link": "https://cloud.google.com/logging/docs/",
-      "usage": "node logs.js --help"
+      "id": "quickstart",
+      "name": "Quickstart",
+      "file": "quickstart.js",
+      "docs_link": "https://cloud.google.com/logging/docs/"
     },
     {
-      "id": "sinks",
-      "name": "Sinks",
-      "file": "sinks.js",
-      "docs_link": "https://cloud.google.com/logging/docs/",
-      "usage": "node sinks.js --help"
+      "id": "setup_explicit",
+      "name": "Explict Auth Setup",
+      "file": "setup_explicit.js",
+      "docs_link": "https://cloud.google.com/logging/docs/"
     }
   ]
 }
