+++ .mailmap
@@ -1,0 +1,6 @@
+Jason Dobry <jdobry@google.com> Jason Dobry <jmdobry@users.noreply.github.com>
+Jason Dobry <jdobry@google.com> Jason Dobry <jason.dobry@gmail.com>
+Luke Sneeringer <lukesneeringer@google.com> Luke Sneeringer <luke@sneeringer.com>
+Stephen Sawchuk <sawchuk@gmail.com> Stephen <stephenplusplus@users.noreply.github.com>
+Stephen Sawchuk <sawchuk@gmail.com> Stephen Sawchuk <stephenplusplus@users.noreply.github.com>
+Stephen Sawchuk <sawchuk@gmail.com> Stephen Sawchuk <stephenplusplusplus@gmail.com>
