+++ CONTRIBUTORS
@@ -8,12 +8,10 @@ Ali Ijaz Sheikh <ofrobots@google.com>
 Bill Prin <waprin@google.com>
 Dave Gramlich <callmehiphop@gmail.com>
 Dave Raffensperger <daveraff@google.com>
-Jason Dobry <jason.dobry@gmail.com>
-Jason Dobry <jmdobry@users.noreply.github.com>
+Jason Dobry <jdobry@google.com>
 Joshua Toenyes <JoshuaToenyes@users.noreply.github.com>
+Luke Sneeringer <lukesneeringer@google.com>
 Michael Gao <fgao@google.com>
 Stephen Sawchuk <sawchuk@gmail.com>
-Stephen Sawchuk <stephenplusplus@users.noreply.github.com>
-Stephen Sawchuk <stephenplusplusplus@gmail.com>
 Tim Swast <swast@google.com>
 circle-ci <circle-ci@circleci.com>
